% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage.R
\name{dml.lm}
\alias{dml.lm}
\title{Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.}
\usage{
dml.lm(
  data,
  y_var,
  d_vars,
  x_vars,
  first_stage_family = c("ols", "rf", "lasso"),
  second_stage_family = c("mr", "sr1", "sr2")
)
}
\arguments{
\item{second_stage_family}{}
}
\description{
Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.
}
