% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-stage.R
\name{dml.lm.wrapper}
\alias{dml.lm.wrapper}
\title{Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.}
\usage{
dml.lm.wrapper(
  formula,
  data,
  first_stage_family = c("ols", "rf", "lasso", "user-defined"),
  predict_fun = NULL,
  second_stage_family = c("mr", "sr1", "sr2"),
  ...
)
}
\arguments{
\item{formula}{placeholder}

\item{data}{placeholder}

\item{first_stage_family}{placeholder}

\item{second_stage_family}{placeholder}

\item{use_default_hyper}{placeholder}
}
\value{
placeholder
}
\description{
Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.
}
\examples{
placeholder
}
