% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{dml_first_stage_lasso}
\alias{dml_first_stage_lasso}
\title{Random forest w/ grid-search possibility
Grid search ranger}
\usage{
dml_first_stage_lasso(
  data,
  x_vars,
  y_var,
  d_vars,
  y_family,
  d_family,
  nfold = 2,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{verbose}{}

\item{node_size}{Title}
}
\description{
Random forest w/ grid-search possibility
Grid search ranger
}
