% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest.R
\name{dml_first_stage_rf}
\alias{dml_first_stage_rf}
\title{Random forest first stage}
\usage{
dml_first_stage_rf(
  data,
  x_vars,
  y_var,
  d_vars,
  verbose = TRUE,
  nfold = 2,
  mtry = NULL,
  num_trees = NULL,
  node_size = NULL,
  error_type = "OOB",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Random forest first stage
}
