% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-functions.R
\name{dml_second_stage}
\alias{dml_second_stage}
\title{Estimates ATE and SE by regressing outcome residuals on treatment residuals}
\usage{
dml_second_stage(y_resid, d_resid)
}
\arguments{
\item{d_resid}{}
}
\description{
Estimates ATE and SE by regressing outcome residuals on treatment residuals
}
