% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-functions.R
\name{dml_second_stage_lasso}
\alias{dml_second_stage_lasso}
\title{This function facilitates running second-stage DML with LASSO}
\usage{
dml_second_stage_lasso(
  formula,
  df,
  nfolds,
  lambda = "1se",
  alpha = 0.05,
  standardize = TRUE,
  post.lasso = FALSE
)
}
\arguments{
\item{post.lasso}{}
}
\description{
This function facilitates running second-stage DML with LASSO
}
