% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-functions.R
\name{dml_second_stage_ridge}
\alias{dml_second_stage_ridge}
\title{This function facilitates running second-stage DML with Ridge}
\usage{
dml_second_stage_ridge(
  formula,
  df,
  nfolds,
  lambda = "lambda.1se",
  alpha = 0.05,
  standardize = TRUE
)
}
\arguments{
\item{standardize}{}
}
\description{
This function facilitates running second-stage DML with Ridge
}
