% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-functions.R
\name{dml_second_stage_ridge_boot}
\alias{dml_second_stage_ridge_boot}
\title{Runs bootstrapped hte_ridge to get coefficients and conf. intervals for them.}
\usage{
dml_second_stage_ridge_boot(
  formula,
  df,
  nfolds,
  lambda = "lambda.min",
  alpha = 0.05,
  nboot = 1000,
  standardize = TRUE,
  strata
)
}
\arguments{
\item{strata}{}
}
\description{
Runs bootstrapped hte_ridge to get coefficients and conf. intervals for them.
}
