% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riplDML-package.R
\docType{package}
\name{linearDML-package}
\alias{linearDML}
\alias{linearDML-package}
\title{linearDML: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yonah Meiselman \email{ymeiselman@ripl.org}

}
\keyword{internal}
