% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dml}
\alias{predict.dml}
\title{Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.}
\usage{
predict.dml(data, y_var, x_vars, family = "ols")
}
\arguments{
\item{second_stage}{}
}
\description{
Estimates treatment effects by regressing outcome residuals on treatment
residuals, where residuals come from predicting outcomes and treatments with
covariates.
}
